Move DeclarationScanner to Knowledge layer
Moves Declaration, DeclarationScanner, and FileDeclarations from
Index to Knowledge. This drains the last deptrac baseline entry:
Repository may now reach these classes through Knowledge, which
it is allowed to depend on.
