+++ phpstan.neon
@@ -17,7 +17,7 @@ parameters:
             allowIn:
                 - src/Parser/ParserService.php
                 - src/Index/NodeAtPosition.php
-                - src/Index/DeclarationScanner.php
+                - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
                 - tests/*
@@ -27,7 +27,7 @@ parameters:
             allowIn:
                 - src/Parser/ParserService.php
                 - src/Index/NodeAtPosition.php
-                - src/Index/DeclarationScanner.php
+                - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
                 - tests/*
@@ -49,7 +49,7 @@ parameters:
             allowIn:
                 - src/Parser/ParserService.php
                 - src/Index/NodeAtPosition.php
-                - src/Index/DeclarationScanner.php
+                - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
                 - tests/*
