+++ src/Knowledge/Declaration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Firehed\PhpLsp\Index;
+namespace Firehed\PhpLsp\Knowledge;
 
 use Firehed\PhpLsp\Domain\QualifiedName;
 use PhpParser\Node;
